app/requirements.txt
alchemy-mock==0.4.3
attrs==19.3.0
bitarray==1.1.0
boto==2.49.0
boto3==1.11.11
botocore==1.14.11
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
crypto==1.4.1
cryptography==2.8
docutils==0.15.2
editdistance==0.5.3
filelock==3.0.12
Flask==1.1.1
future==0.18.2
gensim==3.8.1
gevent==1.4.0
greenlet==0.4.15
grequests==0.4.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
miscreant==0.3.0
mmh3==2.5.1
more-itertools==7.2.0
msgpack==0.6.2
mysql-connector==2.2.9
Naked==0.1.31
numpy==1.18.1
packaging==19.2
pluggy==0.13.1
py==1.8.0
pycparser==2.19
pyparsing==2.4.5
pyprobables==0.3.0
pytest==5.3.1
pytest-mock==1.12.1
python-dateutil==2.8.1
PyYAML==5.3
pyzmq==18.1.1
requests==2.22.0
s3transfer==0.3.2
scipy==1.4.1
shellescape==3.8.1
six==1.13.0
smart-open==1.9.0
SQLAlchemy==1.3.11
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
word2keypress==1.0.16
zerorpc==0.6.3
zipp==0.6.0
zxcvbn==4.4.28
